requirements.txt
google-cloud-scheduler
google-cloud-firestore
ccxt

# See https://cloud.google.com/functions/docs/writing/specifying-dependencies-python
# and https://cloud.google.com/functions/docs/reference/python-system-packages
# Use these dependencies if possible:
# click==6.7
# Flask==1.0.2
# itsdangerous==0.24
# Jinja2==2.10
# MarkupSafe==1.0
# pip==18.0
# requests==2.19.1
# setuptools==40.2.0
# Werkzeug==0.14.1
# wheel==0.31.1
